+++ src/NVMeDriverPatcher/ViewModels/MainViewModel.Settings.cs
@@ -134,7 +134,11 @@ public partial class MainViewModel
         DebouncedSaveSettings();
     }
 
-    partial void OnButtonsEnabledChanged(bool value) => UpdateRecommendedActions();
+    partial void OnButtonsEnabledChanged(bool value)
+    {
+        RefreshMutationActionAvailability();
+        UpdateRecommendedActions();
+    }
     partial void OnApplyEnabledChanged(bool value) => UpdateRecommendedActions();
     partial void OnIsLoadingChanged(bool value) => UpdateRecommendedActions();